\name{phe_as_clogit}
\alias{phe_as_clogit}
\title{PheWAS using conditional logistic regression}

\description{
Called within the phewas_ext() function to perform PheWAS analysis using conditional logistic regression (“clogit”). For more details on arguments/outputs of PheWAS, see the phewas_ext() function.}

\references{
\url{
https://github.com/PheWAS/PheWAS/blob/master/R/phe_as_clogit.R}
}

\note{
Code was not modified from the phe_as_clogit() function in the original PheWAS R package.}
