\name{phe_as_logistf}
\alias{phe_as_logistf}
\title{PheWAS using Firth's logistic regression}

\description{
Called within the phewas_ext() function to perform PheWAS analysis using Firth’s logistic regression (“logistf”). For more details on arguments/outputs of PheWAS, see the phewas_ext() function.}

\value{
Returns a dataframe with information on phenotype, genotype, covariates p value, type, and cohort information.}

\references{
\url{
https://github.com/PheWAS/PheWAS/blob/master/R/phe_as_logistf.R}
}

\note{
Code was not modified from the phe_as_logistf() function in the original PheWAS R package.}
