\name{phe_as_lrt}
\alias{phe_as_lrt}
\title{PheWAS using likelihood ratio test}

\description{
%%Called within the phewas_ext() function to perform a likelihood ratio test (“lrt”). For more details on arguments/outputs of PheWAS, see the phewas_ext() function.}

\references{
\url{
%%https://github.com/PheWAS/PheWAS/blob/master/R/phe_as_lrt.R}
}

\note{
%% Code was not modified from the phe_as_lrt() function in the original PheWAS R package.}
